# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  entities.Campus:
    properties:
      address:
        type: string
      id:
        type: integer
      name:
        type: string
      university:
        type: string
    type: object
...
  entities.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  entities.UpdateDeleteCampusResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  title: VUZ+ API
  version: "1.0"
paths:
...
  /auth/campus:
    put:
      consumes:
      - application/json
      parameters:
      - description: campus data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entities.Campus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.UpdateDeleteCampusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Update campus
      tags:
      - campus
  /auth/campus/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: campus id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.UpdateDeleteCampusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Delete campus
      tags:
      - campus
  /auth/campus/address/{address}:
    get:
      consumes:
      - application/json
      parameters:
      - description: campus address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Campus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get campus by address
      tags:
      - campus
  /auth/campus/all:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Campus'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get all campuses
      tags:
      - campus
  /auth/campus/id/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: campus id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Campus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get campus by id
      tags:
      - campus
  /auth/campus/name/{name}:
    get:
      consumes:
      - application/json
      parameters:
      - description: campus name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Campus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get campus by name
      tags:
      - campus
  /auth/campus/university/{university}:
    get:
      consumes:
      - application/json
      parameters:
      - description: campus university
        in: path
        name: university
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Campus'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get campus by university
      tags:
      - campus
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"